feat:support client and server on the same machine

+++ scripts/tsdbComp/full_cycle_minitest_loading.sh
@@ -159,7 +159,11 @@ fi
 # use different load scripts of db to load data , add supported databases 
 if [ "${FORMAT}" == "timescaledb" ];then
     PGPASSWORD=${DATABASE_PWD} psql -U postgres -h $DATABASE_HOST  -d postgres -c "drop database IF EXISTS  ${DATABASE_NAME} "
-    disk_usage_before=$(set_command "du -s ${TimePath} --exclude="pgsql_tmp" | cut -f 1 " )
+    if [ -d "${TimePath}" ]; then
+        disk_usage_before=$(set_command "du -s ${TimePath} --exclude="pgsql_tmp" | cut -f 1 " )
+    else
+        disk_usage_before=0
+    fi
     echo $disk_usage_before
     echo "BATCH_SIZE":${BATCH_SIZE} "USE_CASE":${USE_CASE} "FORMAT":${FORMAT}  "NUM_WORKER":${NUM_WORKER}  "SCALE":${SCALE}
     RESULT_NAME="${FORMAT}_${USE_CASE}_scale${SCALE}_worker${NUM_WORKER}_batch${BATCH_SIZE}_data.txt"
@@ -220,10 +224,15 @@ if [ "${FORMAT}" == "timescaledb" ];then
     PGPASSWORD=${DATABASE_PWD} psql -U postgres -h $DATABASE_HOST  -d postgres -c "drop database IF EXISTS  ${DATABASE_NAME} "
     sleep 60
 elif [  ${FORMAT} == "influx" ];then
-    run_command "rm -rf ${InfPath}/*
-    systemctl restart influxd
-    sleep 1"
-    disk_usage_before=`set_command "du -s ${InfPath}/data | cut -f 1 " `
+    if [ -d "${InfPath}" ]; then
+        run_command "rm -rf ${InfPath}/*
+            systemctl restart influxd
+            sleep 1"
+        disk_usage_before=`set_command "du -s ${InfPath}/data | cut -f 1 " `
+    else
+        disk_usage_before=0
+    fi
+
     echo "BATCH_SIZE":${BATCH_SIZE} "USE_CASE":${USE_CASE} "FORMAT":${FORMAT}  "NUM_WORKER":${NUM_WORKER}  "SCALE":${SCALE}
     RESULT_NAME="${FORMAT}_${USE_CASE}_scale${SCALE}_worker${NUM_WORKER}_batch${BATCH_SIZE}_data.txt"
     echo `date +%Y_%m%d_%H%M%S`
@@ -274,7 +283,11 @@ elif [  ${FORMAT} == "TDengine" ];then
     sleep 2"
 
     echo "caculte data size"
-    disk_usage_before=`set_command "du -s ${TDPath}/vnode | cut -f 1 " `
+    if [ -d "${TDPath}" ]; then
+        disk_usage_before=`set_command "du -s ${TDPath}/vnode | cut -f 1 " `
+    else
+        disk_usage_before=0
+    fi
     echo "BATCH_SIZE":${BATCH_SIZE} "USE_CASE":${USE_CASE} "FORMAT":${FORMAT}  "NUM_WORKER":${NUM_WORKER}  "SCALE":${SCALE}
     RESULT_NAME="${FORMAT}_${USE_CASE}_scale${SCALE}_worker${NUM_WORKER}_batch${BATCH_SIZE}_data.txt"
     if [ ${SCALE} -ge 100000 ];then
